% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_reduction.R
\name{apply_threshold}
\alias{apply_threshold}
\title{Apply Threshold to Matrix}
\usage{
apply_threshold(mx, threshold, threshold_type, verbose = FALSE)
}
\arguments{
\item{mx}{Matrix to threshold}

\item{threshold}{Threshold specification}

\item{threshold_type}{Type of threshold}

\item{verbose}{Print progress}
}
\value{
List with matrix and threshold_value
}
\description{
Sets values below threshold to zero.
}
\keyword{internal}
