% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_tuning_helpers.R
\name{cache_result}
\alias{cache_result}
\title{Cache Computation Result}
\usage{
cache_result(cache, candidate, cut.off, mx, result)
}
\arguments{
\item{cache}{Cache object.}

\item{candidate}{Parameter candidate.}

\item{cut.off}{Cut-off threshold.}

\item{mx}{Mobility matrix.}

\item{result}{Computation result.}
}
\description{
Stores computation result in cache for future use.
}
\keyword{internal}
