% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_tuning.R
\name{calculate_expected_improvement}
\alias{calculate_expected_improvement}
\title{Calculate Expected Improvement}
\usage{
calculate_expected_improvement(mean, std, best_objective, xi = 0.01)
}
\arguments{
\item{mean}{Predicted means from GP}

\item{std}{Predicted standard deviations from GP}

\item{best_objective}{Current best objective value}

\item{xi}{Exploration parameter}
}
\description{
Computes Expected Improvement acquisition function values.
}
\keyword{internal}
