% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_reduction.R
\name{calculate_ppmi}
\alias{calculate_ppmi}
\title{Calculate Positive Pointwise Mutual Information (PPMI)}
\usage{
calculate_ppmi(mx)
}
\arguments{
\item{mx}{A mobility matrix (core, without totals)}
}
\value{
PPMI-transformed matrix
}
\description{
Transforms a mobility matrix using PPMI to emphasize mobility above chance levels.
}
\keyword{internal}
