% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_tuning.R
\name{cleanup_parallel_cluster}
\alias{cleanup_parallel_cluster}
\title{Clean Up Parallel Cluster}
\usage{
cleanup_parallel_cluster(cluster_setup, verbose = FALSE)
}
\arguments{
\item{cluster_setup}{Cluster setup object returned by setup_parallel_cluster().}

\item{verbose}{Logical indicating whether to show cleanup messages.}
}
\description{
Stops and cleans up a parallel cluster created by setup_parallel_cluster().
}
\keyword{internal}
