% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_tuning.R
\name{compute_adjusted_rand_index}
\alias{compute_adjusted_rand_index}
\title{Compute Adjusted Rand Index}
\usage{
compute_adjusted_rand_index(clustering1, clustering2)
}
\arguments{
\item{clustering1}{First clustering solution.}

\item{clustering2}{Second clustering solution.}
}
\value{
Numeric adjusted rand index.
}
\description{
Computes the adjusted rand index for comparing clustering solutions.
}
\keyword{internal}
