% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_optimization.R
\name{compute_expected_improvement}
\alias{compute_expected_improvement}
\title{Compute Expected Improvement}
\usage{
compute_expected_improvement(gp_model, grid_points, observed_scores)
}
\arguments{
\item{gp_model}{Gaussian Process model.}

\item{grid_points}{Points to evaluate.}

\item{observed_scores}{Current best scores.}
}
\value{
Vector of EI values.
}
\description{
Calculates Expected Improvement acquisition function values.
}
\keyword{internal}
