% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_objective_optimization.R
\name{compute_hypervolume_mc}
\alias{compute_hypervolume_mc}
\title{Monte Carlo Hypervolume Estimation}
\usage{
compute_hypervolume_mc(frontier_objectives, reference_point, n_samples = 10000)
}
\arguments{
\item{frontier_objectives}{Matrix of Pareto frontier points.}

\item{reference_point}{Reference point for hypervolume.}

\item{n_samples}{Number of Monte Carlo samples.}
}
\value{
Estimated hypervolume.
}
\description{
Estimates hypervolume using Monte Carlo sampling for higher dimensions.
}
\keyword{internal}
