% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_optimization.R
\name{compute_probability_improvement}
\alias{compute_probability_improvement}
\title{Compute Probability of Improvement}
\usage{
compute_probability_improvement(gp_model, grid_points, observed_scores)
}
\arguments{
\item{gp_model}{Gaussian Process model.}

\item{grid_points}{Points to evaluate.}

\item{observed_scores}{Current observed scores.}
}
\value{
Vector of PI values.
}
\description{
Calculates Probability of Improvement acquisition function values.
}
\keyword{internal}
