% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{compute_segment_centrality_metrics}
\alias{compute_segment_centrality_metrics}
\title{Compute Segment Centrality Metrics}
\usage{
compute_segment_centrality_metrics(segment_matrix, enhanced_membership)
}
\description{
Compute Segment Centrality Metrics
}
\keyword{internal}
