% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_clustering.R
\name{compute_transition_matrix}
\alias{compute_transition_matrix}
\title{Compute Transition Matrix}
\usage{
compute_transition_matrix(aligned_memberships)
}
\arguments{
\item{aligned_memberships}{List of aligned membership data frames.}
}
\value{
Transition probability matrix.
}
\description{
Internal function to compute segment-to-segment transition probabilities.
}
\keyword{internal}
