% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_tuning_visualization.R
\name{create_convergence_plot}
\alias{create_convergence_plot}
\title{Create Convergence Plot for Bayesian Optimization}
\usage{
create_convergence_plot(tuning_result)
}
\arguments{
\item{tuning_result}{Tuning results object.}
}
\value{
ggplot object.
}
\description{
Shows convergence behavior of Bayesian optimization.
}
\keyword{internal}
