% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_optimization.R
\name{create_latin_hypercube_sample}
\alias{create_latin_hypercube_sample}
\title{Create Latin Hypercube Sample}
\usage{
create_latin_hypercube_sample(bounds, n_samples)
}
\arguments{
\item{bounds}{Vector with min and max values.}

\item{n_samples}{Number of samples to generate.}
}
\value{
Vector of sampled points.
}
\description{
Creates space-filling initial design for Bayesian optimization.
}
\keyword{internal}
