% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_tuning_visualization.R
\name{create_overview_plot}
\alias{create_overview_plot}
\title{Create Overview Plot}
\usage{
create_overview_plot(
  tuning_result,
  candidates,
  quality_scores,
  perf_times,
  stability_scores
)
}
\arguments{
\item{tuning_result}{Tuning results object.}

\item{candidates}{Vector of candidates.}

\item{quality_scores}{Vector of quality scores.}

\item{perf_times}{Vector of performance times.}

\item{stability_scores}{Vector of stability scores.}
}
\value{
ggplot object.
}
\description{
Creates main overview plot showing quality, performance, and stability.
}
\keyword{internal}
