% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_tuning_visualization.R
\name{create_pareto_plot}
\alias{create_pareto_plot}
\title{Create Pareto Frontier Plot}
\usage{
create_pareto_plot(tuning_result)
}
\arguments{
\item{tuning_result}{Tuning results object.}
}
\value{
ggplot object.
}
\description{
Creates plot showing Pareto frontier analysis.
}
\keyword{internal}
