% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_reduction.R
\name{detect_elbow}
\alias{detect_elbow}
\title{Detect Elbow Point in Scree Data}
\usage{
detect_elbow(variance_per_component)
}
\arguments{
\item{variance_per_component}{Variance explained per component}
}
\value{
Integer index of elbow point
}
\description{
Uses the maximum distance method to find the elbow point.
}
\keyword{internal}
