% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_tuning.R
\name{display_progress}
\alias{display_progress}
\title{Display Progress Information}
\usage{
display_progress(tracker, current_progress = NULL, message = NULL)
}
\description{
Display Progress Information
}
\keyword{internal}
