% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph_compatibility.R
\docType{data}
\name{.moneca_env}
\alias{.moneca_env}
\title{igraph Compatibility Layer}
\format{
An object of class \code{environment} of length 7.
}
\usage{
.moneca_env
}
\description{
This file provides compatibility functions to handle both old and new igraph API
ensuring MONECA works with igraph versions from 1.3.0 onwards
}
\keyword{internal}
