% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_tuning.R
\name{estimate_components}
\alias{estimate_components}
\title{Estimate Number of Connected Components}
\usage{
estimate_components(adj_matrix)
}
\arguments{
\item{adj_matrix}{Binary adjacency matrix.}
}
\value{
Estimated number of connected components.
}
\description{
Simple approximation of connected components in a network.
}
\keyword{internal}
