% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff_sensitivity_strength.R
\name{evaluate_segmentation_metrics}
\alias{evaluate_segmentation_metrics}
\title{Evaluate Segmentation Metrics}
\usage{
evaluate_segmentation_metrics(mx, cut.off, small.cell.reduction = 0)
}
\arguments{
\item{mx}{Mobility matrix}

\item{cut.off}{Cut-off value}

\item{small.cell.reduction}{Small cell adjustment}
}
\value{
List of segmentation metrics
}
\description{
Evaluates how a specific cut-off affects the segmentation structure.
}
\keyword{internal}
