% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff_sensitivity_strength.R
\name{evaluate_stability}
\alias{evaluate_stability}
\title{Evaluate Stability via Bootstrap}
\usage{
evaluate_stability(mx, cutoff_values, n_bootstrap = 30)
}
\arguments{
\item{mx}{Mobility matrix}

\item{cutoff_values}{Vector of cut-off values to test}

\item{n_bootstrap}{Number of bootstrap samples}
}
\value{
Vector of stability scores (lower is more stable)
}
\description{
Evaluates the stability of segmentation at different cut-off values
using bootstrap resampling.
}
\keyword{internal}
