% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{extract_named_matrices}
\alias{extract_named_matrices}
\title{Extract Named Matrices}
\usage{
extract_named_matrices(segments, max_level, enhanced_membership)
}
\description{
Extract Named Matrices
}
\keyword{internal}
