% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{find_best_level_with_structure}
\alias{find_best_level_with_structure}
\title{Find Best Level With Structure and Names}
\usage{
find_best_level_with_structure(segments, enhanced_membership, max_level)
}
\description{
Find Best Level With Structure and Names
}
\keyword{internal}
