% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{find_common_suffix}
\alias{find_common_suffix}
\title{Find Common Suffix in Node Names}
\usage{
find_common_suffix(names)
}
\arguments{
\item{names}{Character vector of names}
}
\value{
Character string of common suffix, or empty string if none
}
\description{
Find Common Suffix in Node Names
}
\keyword{internal}
