% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff_rr_analysis.R
\name{find_elbow_point}
\alias{find_elbow_point}
\title{Find Elbow Point in Curve}
\usage{
find_elbow_point(x, y)
}
\arguments{
\item{x}{Numeric vector of x values}

\item{y}{Numeric vector of y values}
}
\value{
x value at the elbow point
}
\description{
Identifies the elbow point in a curve using the perpendicular distance method.
}
\keyword{internal}
