% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff_rr_analysis.R
\name{find_natural_breaks}
\alias{find_natural_breaks}
\title{Find Natural Breaks in RR Distribution}
\usage{
find_natural_breaks(rr_values, n_breaks = 3)
}
\arguments{
\item{rr_values}{Numeric vector of RR values}

\item{n_breaks}{Number of breaks to identify}
}
\value{
Numeric vector of break points
}
\description{
Identifies natural breaks in the RR value distribution using the
Jenks natural breaks optimization method.
}
\keyword{internal}
