% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_functions.R
\name{first.level.summary}
\alias{first.level.summary}
\title{First level summary}
\usage{
first.level.summary(segments,
  small.cell.reduction = segments$small.cell.reduction)
}
\description{
First level summary
}
