% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_tuning.R
\name{fit_gaussian_process}
\alias{fit_gaussian_process}
\title{Fit Gaussian Process Model}
\usage{
fit_gaussian_process(X, y)

fit_gaussian_process(X, y)
}
\arguments{
\item{X}{Matrix of input points (n x 2).}

\item{y}{Vector of function values.}

\item{evaluated_points}{Data frame with scr, cutoff, and objective columns}
}
\value{
GP model object.
}
\description{
Fits a simple Gaussian Process surrogate model to the evaluated points.

Fits a simple GP model using distance-weighted local regression.
}
\keyword{internal}
