% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_tuning.R
\name{generate_cache_key}
\alias{generate_cache_key}
\title{Generate cache key for parameter combinations}
\usage{
generate_cache_key(mx_hash, scr, cutoff, additional = NULL)
}
\description{
Generate cache key for parameter combinations
}
\keyword{internal}
