% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{generate_segment_plot}
\alias{generate_segment_plot}
\title{Generate Segment Network Plot}
\usage{
generate_segment_plot(
  segment_matrix,
  enhanced_membership,
  mobility_metrics,
  centrality_metrics,
  plot_layout,
  plot_node_color,
  node_size_metric
)
}
\description{
Generate Segment Network Plot
}
\keyword{internal}
