% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_tuning.R
\name{get_cache_stats}
\alias{get_cache_stats}
\title{Get Cache Statistics}
\usage{
get_cache_stats()
}
\value{
List containing cache statistics.
}
\description{
Returns information about the current state of evaluation caches.
}
\examples{
\dontrun{
cache_info <- get_cache_stats()
print(cache_info)
}

}
