% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_parallel_switching.R
\name{get_optimal_cores}
\alias{get_optimal_cores}
\title{Get Optimal Number of Cores}
\usage{
get_optimal_cores(
  n_combinations,
  matrix_size,
  system_resources = NULL,
  max_cores = NULL
)
}
\arguments{
\item{n_combinations}{Number of parameter combinations.}

\item{matrix_size}{Size of mobility matrix.}

\item{system_resources}{System resource information.}

\item{max_cores}{Maximum number of cores to consider.}
}
\value{
Integer representing optimal number of cores to use.
}
\description{
Determines the optimal number of cores to use for a given problem,
considering problem characteristics and system resources.
}
\details{
The function optimizes core count by:
- Ensuring sufficient work per core (avoid over-parallelization)
- Considering memory bandwidth limitations
- Balancing parallel overhead vs benefit
- Respecting system resource constraints
}
\examples{
\dontrun{
resources <- detect_system_resources()
optimal_cores <- get_optimal_cores(
  n_combinations = 100,
  matrix_size = 50,
  system_resources = resources
)
}

}
