% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplotting.R
\name{gg.jonas}
\alias{gg.jonas}
\title{ggplots by Jonas}
\usage{
gg.jonas(segmenter, niveau = seq(segmenter$segment.list),
  layout = layout.matrix(segmenter), edges = log(segment.edges(segmenter) +
  1), mode = "directed", vertex.size = "total", vertex.fill = "segment",
  vertex.alpha = 1, vertex.color = "black", vertex.shape = 21,
  show.edges = TRUE, edge.size = 1, edge.alpha = "weight",
  edge.color = "weight", edge.line = "solid", show.text = TRUE,
  text.size = 3, text.color = "black", text.alpha = 1, text.vjust = 1.5,
  show.borders = TRUE, border.size = 1, border.fill = NA,
  border.color = "black", border.alpha = 1, border.padding = 1,
  border.text = TRUE, border.labels = "segments", border.text.size = 4,
  border.text.color = "black", border.text.vjust = -0.2,
  border.text.hjust = 1, midpoints = TRUE, midpoint.arrow = arrow(angle =
  20, length = unit(0.33, "cm"), ends = "last", type = "closed"),
  edge.text = FALSE, edge.text.size = 3, edge.text.alpha = 0.9,
  legend = "side", ...)
}
\arguments{
\item{niveau}{}

\item{layout}{}

\item{edges}{}

\item{mode}{}

\item{vertex.size}{}

\item{vertex.fill}{}

\item{vertex.alpha}{}

\item{vertex.color}{}

\item{vertex.shape}{}

\item{show.edges}{}

\item{edge.size}{}

\item{edge.alpha}{}

\item{edge.color}{}

\item{edge.line}{}

\item{show.text}{}

\item{text.size}{}

\item{text.color}{}

\item{text.alpha}{}

\item{text.vjust}{}

\item{show.borders}{}

\item{border.size}{}

\item{border.fill}{}

\item{border.color}{}

\item{border.alpha}{}

\item{border.padding}{}

\item{border.text}{}

\item{border.labels}{}

\item{border.text.size}{}

\item{border.text.color}{}

\item{border.text.vjust}{}

\item{border.text.hjust}{}

\item{midpoints}{}

\item{midpoint.arrow}{}

\item{edge.text}{}

\item{edge.text.size}{}

\item{edge.text.alpha}{}

\item{legend}{}
}
\description{
Plotting of MONECA objects
}

