% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_objective_optimization.R
\name{\%||\%}
\alias{\%||\%}
\title{%||% Operator for Default Values}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Value to check.}

\item{y}{Default value.}
}
\value{
x if not NULL, otherwise y.
}
\description{
Provides default value if left side is NULL.
}
\keyword{internal}
