% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_tuning_visualization.R
\name{handle_plot_output}
\alias{handle_plot_output}
\title{Handle Plot Output}
\usage{
handle_plot_output(plots, plot_type, save_plot, filename, width, height)
}
\arguments{
\item{plots}{List of plots.}

\item{plot_type}{Type of plot requested.}

\item{save_plot}{Whether to save plot.}

\item{filename}{Filename for saving.}

\item{width}{Plot width.}

\item{height}{Plot height.}
}
\value{
Plot object or list of plots.
}
\description{
Handles the output and saving of plots.
}
\keyword{internal}
