% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{jonas.plot}
\alias{jonas.plot}
\title{Plot Jonas}
\usage{
jonas.plot(segmenter, layout = layout.matrix(segmenter),
  edges = segment.edges(segmenter), mode = "directed",
  niveau = seq(segmenter$segment.list), vertex.size = 5,
  vertex.frame.color = "black", edge.curved = FALSE,
  vertex.color = "grey50", vertex.label.color = "black",
  vertex.label.cex = 0.5, vertex.label.dist = 0.12, edge.arrow.size = 0.1,
  mark.col = NULL, mark.expand = 10, border.col = "black",
  edge.width = 1, edge.color = "black")
}
\description{
Plot the results from a Jonas analysis
}

