% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{jonas}
\alias{jonas}
\title{Jonas}
\usage{
jonas(mat, kliker, cut.off = 1, mode = "symmetric",
  delete.upper.tri = TRUE)
}
\arguments{
\item{mat}{is the mobility table with row and column margins XXXX Er det her sandt? Eller spiser den en relative risk matrice? .}

\item{kliker}{is a list of cliques}

\item{cut.off}{is the minimum weight or relative risk accepted as a network tie}

\item{mode}{defines whether mat is symmetric. If mode is "Mutual" - ties are only created for mutual relations. If mode is "Unmutual", relations are not required to be mutual to result in a tie.}

\item{delete.upper.tri}{defines whether the upper triangle of the matrix is to be deleted. This results in speed gains.}
}
\value{
membership - a numeric vector with discrete group memberships for each row in the matrix.

cliques a list of row indices for each clique
}
\description{
An algorithm for creating discrete groups on the basis of a weighted network and a mobility table
}

