% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{layout.matrix}
\alias{layout.matrix}
\title{Layout matrix}
\usage{
layout.matrix(segments, attraction = c(320, 40, 10, 4, 2),
  level = seq(segments$segment.list), mode = "directed",
  weight.adjustment = 1, start.temp = 20, niter = 10000,
  tie.adjustment = 0.4, ...)
}
\arguments{
\item{segments}{a segment object}

\item{attraction}{the distance between the segment points for each level.}

\item{level}{the included levels}

\item{mode}{the mode}

\item{area.size}{the size of the plot area - see \link{layout.fruchterman.reingold}}
}
\description{
A matrix with the coordinates of the segments
}

