% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_tuning.R
\name{manage_cache_memory}
\alias{manage_cache_memory}
\title{Memory-efficient cache management}
\usage{
manage_cache_memory(max_cache_size = 100)
}
\description{
Memory-efficient cache management
}
\keyword{internal}
