% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mob.seg}
\alias{mob.seg}
\title{Example Segmentation Results}
\format{
A moneca object containing segmentation results.
}
\source{
Generated from mob.mat using moneca().
}
\description{
Pre-computed moneca segmentation results for the mob.mat example data.
This is example data for demonstrating moneca functionality.
}
\keyword{datasets}
\keyword{internal}
