% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{mobility_data_examples}
\alias{mobility_data_examples}
\title{Generate Example Mobility Data}
\description{
This function creates synthetic mobility data for examples and testing.
Use \code{\link{generate_mobility_data}} or \code{\link{generate_example_datasets}}
to create mobility matrices for analysis.
}
\examples{
# Generate basic synthetic data
mobility_data <- generate_mobility_data(n_classes = 5, seed = 123)
seg <- moneca(mobility_data, segment.levels = 2)

# Generate predefined examples
examples <- generate_example_datasets()
simple_seg <- moneca(examples$simple, segment.levels = 2)

}
