% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{moneca}
\alias{moneca}
\title{moneca}
\usage{
moneca(mx = mx, segment.levels = 3, cut.off = 1, mode = "symmetric",
  delete.upper.tri = TRUE, small.cell.reduction = 0)
}
\arguments{
\item{mx}{is a raw mobility table with row and column margins}

\item{segment.levels}{defines the number levels of nested segments to be returned}

\item{cut.off}{is the minimum weight or relative risk accepted as a network tie}

\item{mode}{defines whether mat is symmetric. If mode is "Mutual" - ties are only created for mutual relations. If mode is "Unmutual", relations are not required to be mutual to result in a tie.}

\item{delete.upper.tri}{defines whether the upper triangle of the matrix is to be deleted. This results in speed gains.}
}
\value{
segment.list returns a list of cliques for each level. The row indices correspond to the first level aka. the rows from mx.

mat.list returns a list with a mobility table for each level with the number of rows and columns corresponding to the number of segments for each level
}
\description{
moneca
}
\seealso{
\link{find.segments}, \link{moneca.plot}
}
