% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_functions.R
\name{overordnede.niveau}
\alias{overordnede.niveau}
\title{Overordnende.niveau}
\usage{
overordnede.niveau(segmenter,
  small.cell.reduction = segmenter$small.cell.reduction)
}
\description{
Overordnende.niveau
}

