% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_reduction.R
\name{pearson_residuals}
\alias{pearson_residuals}
\title{Calculate Pearson Residuals}
\usage{
pearson_residuals(mx)
}
\arguments{
\item{mx}{A mobility matrix (core, without totals)}
}
\value{
Matrix of Pearson residuals
}
\description{
Computes Pearson residuals to standardize for marginal distribution effects.
}
\keyword{internal}
