% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_reduction.R
\name{perform_svd}
\alias{perform_svd}
\title{Perform SVD (with optional irlba)}
\usage{
perform_svd(mx, k)
}
\arguments{
\item{mx}{Matrix to decompose}

\item{k}{Number of components}
}
\value{
SVD result list
}
\description{
Perform SVD (with optional irlba)
}
\keyword{internal}
