% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_reduction.R
\name{perform_svd_reduction}
\alias{perform_svd_reduction}
\title{Perform SVD Reduction}
\usage{
perform_svd_reduction(mx, k, verbose = FALSE)
}
\arguments{
\item{mx}{Normalized matrix}

\item{k}{Number of components}

\item{verbose}{Print progress}
}
\value{
Reconstructed matrix
}
\description{
Applies truncated SVD and reconstructs the matrix.
}
\keyword{internal}
