% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff_rr_visualization.R
\name{plot_cutoff_interactive}
\alias{plot_cutoff_interactive}
\title{Create Interactive Cut-off Explorer}
\usage{
plot_cutoff_interactive(
  sensitivity_results,
  metrics = c("strength_ratio", "n_edges", "modularity", "strength_cv")
)
}
\arguments{
\item{sensitivity_results}{Output from cutoff_sensitivity_analysis()}

\item{metrics}{Metrics to include in the interactive plot}
}
\value{
plotly object
}
\description{
Creates an interactive plot for exploring cut-off effects using plotly.
}
