% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff_rr_visualization.R
\name{plot_elbow_curve}
\alias{plot_elbow_curve}
\title{Plot Elbow Curve for Optimal Cut-off}
\usage{
plot_elbow_curve(
  sensitivity_results,
  optimal_cutoff = NULL,
  metric = "strength_ratio"
)
}
\arguments{
\item{sensitivity_results}{Sensitivity analysis results}

\item{optimal_cutoff}{Optimal cut-off value found}

\item{metric}{Metric to use for elbow detection}
}
\value{
ggplot2 object
}
\description{
Visualizes the elbow curve used to identify optimal cut-off.
}
