% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_plotting.R
\name{plot_segment_timeline}
\alias{plot_segment_timeline}
\title{Create Segment Size Timeline}
\usage{
plot_segment_timeline(temporal_data, color_scheme, show_labels)
}
\description{
Internal function to create a timeline of segment sizes.
}
\keyword{internal}
