% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff_rr_visualization.R
\name{plot_strength_comparison}
\alias{plot_strength_comparison}
\title{Plot Strength Comparison Across Cut-offs}
\usage{
plot_strength_comparison(
  mx,
  cutoff_values = c(0.5, 1, 1.5, 2),
  small.cell.reduction = 0
)
}
\arguments{
\item{mx}{Mobility matrix}

\item{cutoff_values}{Vector of cut-off values to compare}

\item{small.cell.reduction}{Small cell adjustment}
}
\value{
ggplot2 object
}
\description{
Creates box plots comparing strength distributions at different cut-off values.
}
