% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_plotting.R
\name{plot_transition_heatmap}
\alias{plot_transition_heatmap}
\title{Create Transition Probability Heatmap}
\usage{
plot_transition_heatmap(temporal_data, color_scheme, show_labels)
}
\description{
Internal function to create a heatmap of transition probabilities.
}
\keyword{internal}
