% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_tuning.R
\name{predict_gaussian_process}
\alias{predict_gaussian_process}
\title{Predict Using Gaussian Process Model}
\usage{
predict_gaussian_process(model, X_new)

predict_gaussian_process(model, X_new)
}
\arguments{
\item{model}{GP model object.}

\item{X_new}{Matrix of new input points.}

\item{gp_model}{Fitted GP model from fit_gaussian_process}
}
\value{
List with mean and variance predictions.
}
\description{
Makes predictions at new points using the GP surrogate.

Makes predictions using the fitted GP model.
}
\keyword{internal}
