% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff_sensitivity_strength.R
\name{print.cutoff_sensitivity}
\alias{print.cutoff_sensitivity}
\title{Print Sensitivity Analysis Summary}
\usage{
\method{print}{cutoff_sensitivity}(x, n = 10, ...)
}
\arguments{
\item{x}{Sensitivity analysis object}

\item{n}{Number of rows to display}

\item{...}{Additional arguments}
}
\description{
Print Sensitivity Analysis Summary
}
