% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_functions.R
\name{print.moneca}
\alias{print.moneca}
\title{Print moneca descriptives}
\usage{
\method{print}{moneca}(segments,
  small.cell.reduction = segments$small.cell.reduction)
}
\description{
Print moneca descriptives
}

