% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_diagnostics.R
\name{print.moneca_parallel_benchmark}
\alias{print.moneca_parallel_benchmark}
\title{Print Method for Parallel Benchmark Results}
\usage{
\method{print}{moneca_parallel_benchmark}(x, ...)
}
\arguments{
\item{x}{Object of class "moneca_parallel_benchmark".}

\item{...}{Additional arguments (unused).}
}
\description{
Print Method for Parallel Benchmark Results
}
