% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_diagnostics.R
\name{print.moneca_parallel_diagnostic}
\alias{print.moneca_parallel_diagnostic}
\title{Print Method for Parallel Diagnostic Results}
\usage{
\method{print}{moneca_parallel_diagnostic}(x, ...)
}
\arguments{
\item{x}{Object of class "moneca_parallel_diagnostic".}

\item{...}{Additional arguments (unused).}
}
\description{
Print Method for Parallel Diagnostic Results
}
