% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_diagnostics.R
\name{print.moneca_resource_monitor}
\alias{print.moneca_resource_monitor}
\title{Print Method for Resource Monitor Results}
\usage{
\method{print}{moneca_resource_monitor}(x, ...)
}
\arguments{
\item{x}{Object of class "moneca_resource_monitor".}

\item{...}{Additional arguments (unused).}
}
\description{
Print Method for Resource Monitor Results
}
