% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moneca_segments.R
\name{print.moneca_segments}
\alias{print.moneca_segments}
\title{Print Method for moneca_segments Objects}
\usage{
\method{print}{moneca_segments}(x, ...)
}
\arguments{
\item{x}{A \code{moneca_segments} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Method for moneca_segments Objects
}
