% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_clustering.R
\name{print.moneca_temporal}
\alias{print.moneca_temporal}
\title{Print Method for moneca_temporal Objects}
\usage{
\method{print}{moneca_temporal}(x, ...)
}
\arguments{
\item{x}{A moneca_temporal object.}

\item{...}{Additional arguments (unused).}
}
\description{
Print Method for moneca_temporal Objects
}
