% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff_sensitivity_strength.R
\name{print.optimal_cutoff}
\alias{print.optimal_cutoff}
\title{Print Optimal Cut-off Result}
\usage{
\method{print}{optimal_cutoff}(x, ...)
}
\arguments{
\item{x}{Optimal cut-off object}

\item{...}{Additional arguments}
}
\description{
Print Optimal Cut-off Result
}
