% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff_rr_analysis.R
\name{print.rr_distribution}
\alias{print.rr_distribution}
\title{Print RR Distribution Analysis}
\usage{
\method{print}{rr_distribution}(x, ...)
}
\arguments{
\item{x}{RR distribution object}

\item{...}{Additional arguments}
}
\description{
Print RR Distribution Analysis
}
