% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_stability.R
\name{print.temporal_stability}
\alias{print.temporal_stability}
\title{Print Method for temporal_stability Objects}
\usage{
\method{print}{temporal_stability}(x, n_top = 5, ...)
}
\arguments{
\item{x}{A temporal_stability object.}

\item{n_top}{Number of top segments to show. Default is 5.}

\item{...}{Additional arguments (unused).}
}
\description{
Print Method for temporal_stability Objects
}
