% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_tuning.R
\name{report_worker_progress}
\alias{report_worker_progress}
\title{Worker Progress Reporter for Parallel Processing}
\usage{
report_worker_progress(tracker_info, increment = 1, message = NULL)
}
\description{
Worker Progress Reporter for Parallel Processing
}
\keyword{internal}
