% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{segment.colors}
\alias{segment.colors}
\title{Segment colors}
\usage{
segment.colors(segmenter)
}
\arguments{
\item{segmenter}{}
}
\description{
Creates a grey scale for the segments
}

