% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{segment.edges}
\alias{segment.edges}
\title{Segment edges}
\usage{
segment.edges(segmenter, cut.off = 1, mode = "directed",
  niveau = seq(segmenter$segment.list),
  segment.reduction = seq(segmenter$segment.list), method = "all",
  top = 3, diagonal = NULL, small.cell.reduction = 0)
}
\description{
The coordinates of the edges
}

