% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_moneca.R
\name{segment.matrix.parallel}
\alias{segment.matrix.parallel}
\title{Parallel Segment Matrix}
\usage{
segment.matrix.parallel(mx, segments, n.cores = NULL)
}
\arguments{
\item{mx}{Mobility matrix}

\item{segments}{List of segment memberships}

\item{n.cores}{Number of cores to use}
}
\value{
Aggregated mobility matrix
}
\description{
Parallel version of segment.matrix() that aggregates mobility matrices.
}
