% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_moneca.R
\name{setup_parallel_backend}
\alias{setup_parallel_backend}
\title{Setup Parallel Backend}
\usage{
setup_parallel_backend(n.cores, backend, data_size)
}
\arguments{
\item{n.cores}{Requested number of cores}

\item{backend}{Requested backend type}

\item{data_size}{Size of the data}
}
\value{
Configuration list
}
\description{
Configure the parallel processing backend based on system and data characteristics.
}
\keyword{internal}
