% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplotting.R
\name{stair.plot}
\alias{stair.plot}
\title{Stair plot}
\usage{
stair.plot(segmenter, niveau = seq(segmenter$segment.list),
  layout = layout.matrix(segmenter), edges = segment.edges(segmenter,
  cut.off = 1, method = "all", segment.reduction = 0, niveau = 1),
  mode = "directed", vertex.size = "total", vertex.alpha = 1,
  vertex.color = "black", vertex.shape = 21, show.edges = TRUE,
  edge.size = 0.5, edge.alpha = "weight", edge.color = "black",
  edge.line = "solid", show.text = FALSE, text.size = 3,
  text.color = "black", text.alpha = 1, text.vjust = 1.5,
  show.borders = TRUE, border.size = 1, border.fill = NA,
  border.color = "black", border.alpha = 1, border.padding = 1,
  border.text = TRUE, border.labels = "segments", border.text.size = 4,
  border.text.color = "black", border.text.vjust = -0.2,
  border.text.hjust = 1, midpoints = TRUE, midpoint.arrow = arrow(angle =
  20, length = unit(0.33, "cm"), ends = "last", type = "closed"),
  edge.text = FALSE, edge.text.size = 3, edge.text.alpha = 0.9,
  legend = "side", level.title = "Level", ...)
}
\arguments{
\item{niveau}{}

\item{layout}{}

\item{edges}{}

\item{mode}{}

\item{vertex.size}{}

\item{vertex.alpha}{}

\item{vertex.color}{}

\item{vertex.shape}{}

\item{show.edges}{}

\item{edge.size}{}

\item{edge.alpha}{}

\item{edge.color}{}

\item{edge.line}{}

\item{show.text}{}

\item{text.size}{}

\item{text.color}{}

\item{text.alpha}{}

\item{text.vjust}{}

\item{show.borders}{}

\item{border.size}{}

\item{border.fill}{}

\item{border.color}{}

\item{border.alpha}{}

\item{border.padding}{}

\item{border.text}{}

\item{border.labels}{}

\item{border.text.size}{}

\item{border.text.color}{}

\item{border.text.vjust}{}

\item{border.text.hjust}{}

\item{midpoints}{}

\item{midpoint.arrow}{}

\item{edge.text}{}

\item{edge.text.size}{}

\item{edge.text.alpha}{}

\item{legend}{}

\item{level.title}{}
}
\value{
a list of ggplot2 objects
}
\description{
Plots the change in segmentation for each level in a MONECA analysis
}
\examples{
data(occupations)
stair.plot(mob.seg)[[2]]
}

