% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_moneca.R
\name{update_segments}
\alias{update_segments}
\title{Update Segments}
\usage{
update_segments(old_segments, new_clusters)
}
\arguments{
\item{old_segments}{Previous segment list}

\item{new_clusters}{New clustering}
}
\value{
Updated segment list
}
\description{
Helper function to update segment list based on new clustering.
}
\keyword{internal}
