% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{weight.matrix}
\alias{weight.matrix}
\title{Weight.matrix}
\usage{
weight.matrix(mx, cut.off = 1, symmetric = TRUE, diagonal = NULL,
  small.cell.reduction = 0)
}
\arguments{
\item{mx}{}

\item{cut.off}{is the minimum weight allowed for a tie between nodes}

\item{symmetric}{defines whether the matrix is forced into symmetry}
}
\value{
a matrix of relative risks
}
\description{
Creates a matrix of weights for the \link{find.segments} function.
}
