% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplotting.R
\name{ego.plot}
\alias{ego.plot}
\title{Ego plots}
\usage{
ego.plot(segmenter, mxa.b, id = 1, lay = layout.matrix(segmenter),
  edge.color = "black", edge.size = 0.8, border.padding = 1,
  title.line = TRUE, ...)
}
\arguments{
\item{segmenter}{}

\item{mxa.b}{}

\item{id}{}

\item{lay}{}

\item{edge.color}{}

\item{edge.size}{}

\item{border.padding}{}

\item{title.line}{}

\item{...}{}
}
\description{
Plots of the connections and movements from a single node
}

