% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/descriptive_functions.R
\name{first.level.summary}
\alias{first.level.summary}
\title{First level summary}
\usage{
first.level.summary(segmenter,
  small.cell.reduction = segmenter$small.cell.reduction)
}
\description{
First level summary
}

