% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_objective_optimization.R
\name{grapes-or-or-grapes}
\alias{grapes-or-or-grapes}
\alias{\%||\%}
\title{Null-coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Value to check for NULL.}

\item{y}{Default value to return if \code{x} is NULL.}
}
\value{
\code{x} if not NULL, otherwise \code{y}.
}
\description{
Provides default value if left side is NULL. This is a convenience operator
that returns the left operand if it is not NULL, otherwise returns the right operand.
This operator is commonly used for setting default values in function arguments.
}
\examples{
\dontshow{
# Internal function examples
value1 <- NULL
value2 <- "default"
result <- value1 \%||\% value2  # returns "default"

value3 <- "actual"
result2 <- value3 \%||\% "default"  # returns "actual"
}
}
\keyword{internal}
