% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{layout.matrix}
\alias{layout.matrix}
\title{Layout matrix}
\usage{
layout.matrix(segmenter, attraction = c(200, 100, 15, 5, 3),
  area.size = 2e+05, niveau = seq(segmenter$segment.list),
  mode = "directed", weight.adjustment = 1, ...)
}
\arguments{
\item{segmenter}{a segment object}

\item{attraction}{the distance between the segment points for each level.}

\item{area.size}{the size of the plot area - see \link{layout.fruchterman.reingold}}

\item{niveau}{the included levels}

\item{mode}{the mode}
}
\description{
A matrix with the coordinates of the segments
}

