% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{segment.edges}
\alias{segment.edges}
\title{Segment edges}
\usage{
segment.edges(segments, cut.off = 1, mode = "directed",
  level = seq(segments$segment.list),
  segment.reduction = seq(segments$segment.list), method = "all", top = 3,
  diagonal = NULL, small.cell.reduction = 0)
}
\description{
The coordinates of the edges
}

