% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{segment.membership}
\alias{segment.membership}
\title{Segment membership}
\usage{
segment.membership(segments, level = seq(segments$segment.list))
}
\description{
A dataframe with the segment membership for each category
}

